* I AM AWESOME * made drag and drop work with filter selector so that now one can assign favorites easily * had to update the database schema for this to work
git-svn-id: https://svn.code.sf.net/p/penguintv/code/trunk@126 a0c6dda6-8710-0410-9408-9ac0bdba4c84

+++ penguintv/penguintv.py
@@ -443,10 +443,13 @@ class PenguinTVApp:
 		new_set = sets.Set(new_tags)
 		removed_tags = list(old_set.difference(new_set))
 		added_tags = list(new_set.difference(old_set))
-		for tag in removed_tags:
-			self.db.remove_tag_from_feed(feed_id, tag)
-		for tag in added_tags:
-			self.db.add_tag_for_feed(feed_id, tag)	
+		if len(removed_tags) == 1 and len(added_tags) == 1: #assume renaming
+			self.db.rename_tag(removed_tags[0], added_tags[0])
+		else:
+			for tag in removed_tags:
+				self.db.remove_tag_from_feed(feed_id, tag)
+			for tag in added_tags:
+				self.db.add_tag_for_feed(feed_id, tag)	
 		if removed_tags or added_tags:
 			self.feed_list_view.set_selected(feed_id)
 		self.main_window.update_filters()
